requirements.txt
discord.py>=2.3.0
privy-python-sdk>=0.1.0
